Organize model summary utilities (#13893)
