+++ src/pytorch_lightning/utilities/model_summary/model_summary_deepspeed.py
@@ -20,7 +20,7 @@ import torch
 from torch.nn import Parameter
 
 from pytorch_lightning.utilities.imports import _RequirementAvailable
-from pytorch_lightning.utilities.model_summary import (
+from pytorch_lightning.utilities.model_summary.model_summary import (
     _is_lazy_weight_tensor,
     get_human_readable_count,
     LayerSummary,
